Disable FileWatcher tests on Linux until the inotify version is written

+++ test/unittests/t_file_watcher.cc
@@ -56,20 +56,23 @@ TEST_F(T_FileWatcher, kModifiedEvent) {
 
   UniquePtr<file_watcher::FileWatcher> watcher(platform_file_watcher());
 
-  TestEventHandler* hd(new TestEventHandler(&counters_, channel_.weak_ref()));
-  watcher->RegisterHandler(watched_file_name, hd);
+  // TODO(radu): Remove this check when the Linux inotify version is added
+  if (watcher.IsValid()) {
+    TestEventHandler* hd(new TestEventHandler(&counters_, channel_.weak_ref()));
+    watcher->RegisterHandler(watched_file_name, hd);
 
-  EXPECT_TRUE(watcher->Start());
+    EXPECT_TRUE(watcher->Start());
 
-  SafeWriteToFile("test", watched_file_name, 0600);
+    SafeWriteToFile("test", watched_file_name, 0600);
 
-  channel_->Dequeue();
+    channel_->Dequeue();
 
-  Counters::const_iterator it_mod = counters_.find(file_watcher::kModified);
-  const int num_modifications = it_mod->second;
-  EXPECT_EQ(1, num_modifications);
+    Counters::const_iterator it_mod = counters_.find(file_watcher::kModified);
+    const int num_modifications = it_mod->second;
+    EXPECT_EQ(1, num_modifications);
 
-  watcher->Stop();
+    watcher->Stop();
+  }
 }
 
 TEST_F(T_FileWatcher, kDeletedEvent) {
@@ -79,20 +82,23 @@ TEST_F(T_FileWatcher, kDeletedEvent) {
 
   UniquePtr<file_watcher::FileWatcher> watcher(platform_file_watcher());
 
-  TestEventHandler* hd(new TestEventHandler(&counters_, channel_.weak_ref()));
-  watcher->RegisterHandler(watched_file_name, hd);
+  // TODO(radu): Remove this check when the Linux inotify version is added
+  if (watcher.IsValid()) {
+    TestEventHandler* hd(new TestEventHandler(&counters_, channel_.weak_ref()));
+    watcher->RegisterHandler(watched_file_name, hd);
 
-  EXPECT_TRUE(watcher->Start());
+    EXPECT_TRUE(watcher->Start());
 
-  unlink(watched_file_name.c_str());
+    unlink(watched_file_name.c_str());
 
-  channel_->Dequeue();
+    channel_->Dequeue();
 
-  Counters::const_iterator it_del = counters_.find(file_watcher::kDeleted);
-  const int num_deletions = it_del->second;
-  EXPECT_EQ(1, num_deletions);
+    Counters::const_iterator it_del = counters_.find(file_watcher::kDeleted);
+    const int num_deletions = it_del->second;
+    EXPECT_EQ(1, num_deletions);
 
-  watcher->Stop();
+    watcher->Stop();
+  }
 }
 
 TEST_F(T_FileWatcher, kModifiedThenDeletedEvent) {
@@ -102,29 +108,32 @@ TEST_F(T_FileWatcher, kModifiedThenDeletedEvent) {
 
   UniquePtr<file_watcher::FileWatcher> watcher(platform_file_watcher());
 
-  TestEventHandler* hd(new TestEventHandler(&counters_, channel_.weak_ref()));
-  hd->clear_ = false;
-  watcher->RegisterHandler(watched_file_name, hd);
+  // TODO(radu): Remove this check when the Linux inotify version is added
+  if (watcher.IsValid()) {
+    TestEventHandler* hd(new TestEventHandler(&counters_, channel_.weak_ref()));
+    hd->clear_ = false;
+    watcher->RegisterHandler(watched_file_name, hd);
 
-  EXPECT_TRUE(watcher->Start());
+    EXPECT_TRUE(watcher->Start());
 
-  SafeWriteToFile("test", watched_file_name, 0600);
+    SafeWriteToFile("test", watched_file_name, 0600);
 
-  channel_->Dequeue();
+    channel_->Dequeue();
 
-  hd->clear_ = true;
+    hd->clear_ = true;
 
-  Counters::const_iterator it_mod = counters_.find(file_watcher::kModified);
-  const int num_modifications = it_mod->second;
-  EXPECT_EQ(1, num_modifications);
+    Counters::const_iterator it_mod = counters_.find(file_watcher::kModified);
+    const int num_modifications = it_mod->second;
+    EXPECT_EQ(1, num_modifications);
 
-  unlink(watched_file_name.c_str());
+    unlink(watched_file_name.c_str());
 
-  channel_->Dequeue();
+    channel_->Dequeue();
 
-  Counters::const_iterator it_del = counters_.find(file_watcher::kDeleted);
-  const int num_deletions = it_del->second;
-  EXPECT_EQ(1, num_deletions);
+    Counters::const_iterator it_del = counters_.find(file_watcher::kDeleted);
+    const int num_deletions = it_del->second;
+    EXPECT_EQ(1, num_deletions);
 
-  watcher->Stop();
+    watcher->Stop();
+  }
 }
